Separate stats functions from Profiles.
